Issue #2397597: Split short/long question to separated project.
