--- modules/question_types/long_answer/long_answer.info
@@ -1,9 +1,0 @@
-name = Long Answer
-package = Quiz Question
-description = This provides long answer (essay, multi-paragraph) question types for use by the Quiz module.
-dependencies[] = quizz
-dependencies[] = quizz_question
-core = 7.x
-
-files[] = long_answer.classes.inc
-files[] = tests/LongAnswerUnitTest.test
