--- modules/question_types/short_answer/short_answer.info
@@ -1,9 +1,0 @@
-name = Short Answer
-package = Quiz Question
-description = Short answer question type for Quiz.
-core = 7.x
-dependencies[] = quizz
-dependencies[] = quizz_question
-
-files[] = short_answer.classes.inc
-files[] = tests/ShortAnswerTestCase.test
