+++ quizz_text/quizz_text.info
@@ -1,0 +1,9 @@
+name = Quizz text
+package = Quiz Question
+description = This provides short and long answer (essay, multi-paragraph) question types for use by the Quizz module.
+dependencies[] = quizz
+dependencies[] = quizz_question
+core = 7.x
+
+files[] = tests/LongAnswerTestCase.test
+files[] = tests/ShortAnswerTestCase.test
