+++ quizz_text/theme/long_answer.theme.inc
