+++ quizz_text/theme/short_answer.theme.inc
